refactor: kafka event 디렉터리 분리
