+++ src/main/java/org/phoenix/planet/consumer/EcoCarEnterDetectedConsumer.java
@@ -8,8 +8,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.phoenix.planet.constant.KafkaTopic;
 import org.phoenix.planet.dto.car.response.MemberCarResponse;
-import org.phoenix.planet.dto.car_access.raw.EcoCarEnterEvent;
 import org.phoenix.planet.dto.eco_stock.raw.EcoStock;
+import org.phoenix.planet.event.EcoCarEnterEvent;
 import org.phoenix.planet.service.car.MemberCarService;
 import org.phoenix.planet.service.eco_stock.EcoStockIssueService;
 import org.phoenix.planet.service.eco_stock.EcoStockService;
